lib/message_passing: Fix complex value conversion finding
Misra rule 5-0-3.

issue: SWP-218373
GIT_ORIGIN_SPP_REV_ID: ab1b8d57ffe481063f79292dbbd200c667fe29b0

+++ score/message_passing/client_connection.cpp
@@ -67,7 +67,7 @@ ClientConnection::ClientConnection(std::shared_ptr<ISharedResourceEngine> engine
     // TODO: separate max_queued_sends, max_async_replies, and maybe queued SendWaitReply
     for (auto& send_command : send_storage_)
     {
-        send_command.message.reserve(max_send_size_);
+        send_command.message.reserve(static_cast<std::size_t>(max_send_size_));
     }
     send_pool_.assign(send_storage_.begin(), send_storage_.end());
 }
